dapp-update: migarated isCorrectHederaNetwork from /api/wallet to /helpers
