+++ system-contract-dapp-playground/src/utils/helpers.ts
@@ -18,7 +18,9 @@
  *
  */
 
-import { PROTECTED_ROUTES } from './constants';
+import { ethers } from 'ethers';
+import { getCurrentChainId } from '@/api/wallet';
+import { HEDERA_NETWORKS, PROTECTED_ROUTES } from './constants';
 
 /**
  * @dev validating if a route is protected
@@ -27,4 +29,23 @@ import { PROTECTED_ROUTES } from './constants';
  */
 export const isProtectedRoute = (pathname: string) => {
   return PROTECTED_ROUTES.includes(pathname);
+};
+
+/**
+ * @dev Handles checking if the connected network is the expected network (i.e. HEDERA_TESTNET, HEDERA_PREVIEWNET, HEDERA_LOCALNET, HEDERA_MAINNET)
+ *
+ * @params walletProvider: ethers.BrowserProvider
+ *
+ * @returns bool
+ */
+export const isCorrectHederaNetwork = async (walletProvider: ethers.BrowserProvider) => {
+  // get current chainId
+  const currentChainId = (await getCurrentChainId(walletProvider)).currentChainId as string;
+
+  return (
+    currentChainId === HEDERA_NETWORKS.mainnet.chainIdHex ||
+    currentChainId === HEDERA_NETWORKS.testnet.chainIdHex ||
+    currentChainId === HEDERA_NETWORKS.previewnet.chainIdHex ||
+    currentChainId === HEDERA_NETWORKS.localnet.chainIdHex
+  );
 };
